Move fixtures from workflow test to utils __fixtures__
