+++ packages/giselle/src/engine/utils/__fixtures__/workspace1.json
