+++ packages/giselle/src/engine/utils/__fixtures__/workspace2.json
