+++ packages/giselle/src/engine/utils/__fixtures__/workspace3.json
