+++ packages/giselle/src/engine/utils/__fixtures__/workspace4.json
